@@ -11,61 +17,67 @@
   },
   "dependencies": {
     "async": "^1.2.1",
-    "babel-core": "^6.9.1",
-    "babel-loader": "^6.2.4",
+    "babel-core": "^6.13.2",
+    "babel-eslint": "^7.1.1",
+    "babel-loader": "^6.2.5",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-plugin-transform-object-assign": "^6.8.0",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.5.0",
+    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-react": "^6.11.1",
     "babel-preset-react-hmre": "^1.1.1",
     "babel-preset-stage-0": "^6.5.0",
+    "babel-traverse": "^6.18.0",
+    "babylon": "^6.14.1",
     "boom": "^2.7.2",
+    "chalk": "^1.1.3",
     "cjsx-loader": "^3.0.0",
     "coffee-loader": "^0.7.2",
     "coffee-script": "^1.9.3",
     "commander": "^2.9.0",
-    "css-loader": "^0.23.1",
+    "css-loader": "^0.25.0",
     "debug": "^2.2.0",
+    "detect-port": "^1.0.0",
     "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.8.5",
+    "file-loader": "^0.9.0",
     "front-matter": "^2.1.0",
     "fs-extra": "^0.30.0",
-    "glob": "^7.0.4",
+    "glob": "^7.0.5",
+    "global": "^4.3.1",
     "hapi": "^8.5.1",
     "hapi-webpack-plugin": "^1.3.0",
-    "highlight.js": "^9.4.0",
+    "highlight.js": "^9.6.0",
     "history": "^2.1.2",
     "html-frontmatter": "^1.6.0",
+    "image-webpack-loader": "^2.0.0",
     "invariant": "^2.2.1",
     "json-loader": "^0.5.2",
     "json5": "^0.5.0",
     "less": "^2.7.1",
     "less-loader": "^2.2.0",
     "loader-utils": "^0.2.15",
-    "lodash": "^4.13.1",
-    "markdown-it": "^6.0.5",
-    "moment": "^2.11.2",
+    "lodash": "^4.15.0",
+    "markdown-it": "^7.0.1",
+    "moment": "^2.14.1",
     "negotiator": "^0.6.1",
     "node-cjsx": "^1.0.0",
     "node-libs-browser": "^1.0.0",
-    "node-sass": "^3.7.0",
+    "node-sass": "^3.8.0",
     "null-loader": "^0.1.1",
     "object-assign": "^4.1.0",
     "opn": "^4.0.2",
     "parse-filepath": "^1.0.1",
     "postcss-browser-reporter": "^0.5.0",
-    "postcss-cssnext": "^2.6.0",
+    "postcss-cssnext": "^2.7.0",
     "postcss-import": "^8.1.2",
-    "postcss-loader": "^0.9.1",
-    "postcss-reporter": "^1.3.3",
-    "process-finder": "^1.0.0",
+    "postcss-loader": "^0.13.0",
+    "postcss-reporter": "^1.4.1",
     "raw-loader": "^0.5.1",
-    "react": "^15.1.0",
+    "react": "^15.4.1",
     "react-document-title": "^2.0.1",
-    "react-dom": "^15.1.0",
+    "react-dom": "^15.4.1",
     "react-hot-loader": "^1.3.0",
-    "react-router": "^2.4.1",
-    "react-router-scroll": "^0.2.0",
+    "react-router": "^2.6.1",
+    "react-router-scroll": "^0.3.1",
     "sass-loader": "^3.1.2",
     "slash": "^1.0.0",
     "static-site-generator-webpack-plugin": "^2.1.0",
@@ -73,29 +85,37 @@
     "toml": "^2.2.2",
     "toml-loader": "^1.0.0",
     "tracer": "^0.8.3",
-    "typography": "^0.9.1",
+    "ts-loader": "^1.2.2",
     "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
+    "webpack": "^1.13.2",
     "webpack-configurator": "^0.3.0",
-    "webpack-hot-middleware": "^2.7.1",
+    "webpack-hot-middleware": "^2.12.2",
     "webpack-require": "0.0.16",
-    "yaml-loader": "^0.2.0"
+    "yaml-loader": "^0.4.0"
   },
   "devDependencies": {
-    "ava": "^0.15.2",
+    "ava": "^0.16.0",
     "ava-http": "^0.2.1",
-    "babel-cli": "^6.9.0",
-    "babel-eslint": "^6.0.2",
-    "babel-register": "^6.9.0",
-    "bluebird": "^3.4.0",
+    "babel-cli": "^6.11.4",
+    "babel-eslint": "^6.1.2",
+    "babel-plugin-lodash": "^3.2.8",
+    "babel-plugin-transform-flow-strip-types": "^6.8.0",
+    "babel-register": "^6.11.6",
+    "bluebird": "^3.4.1",
     "cheerio": "^0.20.0",
-    "eslint": "^2.11.1",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-ava": "^2.5.0",
-    "eslint-plugin-import": "^1.8.1",
-    "eslint-plugin-jsx-a11y": "^1.5.3",
-    "eslint-plugin-react": "^5.1.1",
-    "nyc": "^6.6.1"
+    "eslint": "3.x",
+    "eslint-config-airbnb": "^13.0.0",
+    "eslint-plugin-ava": "^4.0.0",
+    "eslint-plugin-flow-vars": "^0.5.0",
+    "eslint-plugin-flowtype": "^2.7.1",
+    "eslint-plugin-import": "2.2.x",
+    "eslint-plugin-jsx-a11y": "2.2.3",
+    "eslint-plugin-react": "6.8.x",
+    "flow-bin": "^0.30.0",
+    "iflow-debug": "^1.0.15",
+    "iflow-lodash": "^1.1.23",
+    "iflow-react-router": "^1.1.17",
+    "nyc": "^7.0.0"
   },
   "engines": {
     "node": ">0.12.0"
