@@ -1,8 +1,8 @@
 const program = require('commander')
 const packageJson = require('../../package.json')
-const _ = require('lodash')
 const fs = require('fs-extra')
 const path = require('path')
+const _ = require('lodash')
 
 // Copy our load-context function to root of site in a dot file.
 const gatsbyFile = `${__dirname}/../utils/load-context.js`
@@ -24,7 +24,7 @@ program.command('develop')
   .description('Start development server. Watches files and rebuilds and hot reloads if something changes') // eslint-disable-line max-len
   .option('-H, --host <url>',
           `Set host. Defaults to ${defaultHost}`,
-          defaultHost
+          defaultHost,
          )
   .option('-p, --port <port>', 'Set port. Defaults to 8000', '8000')
   .option('-o, --open', 'Open the site in your browser for you.')
@@ -41,6 +41,9 @@ program.command('build')
   .description('Build a Gatsby project.')
   .option('--prefix-links', 'Build site with links prefixed (set prefix in your config).')
   .action((command) => {
+    // Set NODE_ENV to 'production'
+    process.env.NODE_ENV = 'production'
+
     const build = require('../utils/build')
     const p = {
       ...command,
@@ -59,7 +62,7 @@ program.command('serve-build')
   .description('Serve built site.')
   .option('-H, --host <url>',
           `Set host. Defaults to ${defaultHost}`,
-          defaultHost
+          defaultHost,
          )
   .option('-p, --port <port>', 'Set port. Defaults to 8000', '8000')
   .option('-o, --open', 'Open the site in your browser for you.')
