@@ -1,19 +1,33 @@
-import filter from 'lodash/filter'
-import sortBy from 'lodash/sortBy'
-import last from 'lodash/last'
+/* @flow weak */
+import _ from 'lodash'
 import { prefixLink } from './gatsby-helpers'
+
+function requireComponent (moduleReq) {
+  // Minor differences in `require` behavior across languages
+  // are handled by this wrapper.
+  //
+  // If `moduleReq` is a typescript module (versus javascript/coffeescript)
+  // an object with a `default` property is returned rather than the
+  // default export.
+  let module = moduleReq
+  if (module.default) {
+    module = module.default
+  }
+  return module
+}
 
 module.exports = (files, pagesReq) => {
   // Remove files that start with an underscore as this indicates
   // the file shouldn't be turned into a page.
-  const pages = filter(files, (file) => file.file.name.slice(0, 1) !== '_')
-  const templates = filter(files, (file) =>
+  const pages = _.filter(files, file => file.file.name.slice(0, 1) !== '_')
+  const templates = _.filter(files, file =>
+    // eslint-disable-next-line comma-dangle
     file.file.name === '_template'
   )
 
   const routes = {
     path: prefixLink('/'),
-    component: require('pages/_template'),
+    component: requireComponent(require('pages/_template')),
     childRoutes: [],
     indexRoute: {},
     pages,
@@ -33,28 +47,30 @@ module.exports = (files, pagesReq) => {
   // 3. For each index file paired with a template, create a default route
   // 4. Create normal routes for each remaining file under the appropriate
   // template
-  const templatesWithoutRoot = filter(files, (file) =>
+  const templatesWithoutRoot = _.filter(files, file =>
+    // eslint-disable-next-line comma-dangle
     file.file.name === '_template' && file.file.dirname !== ''
   )
 
   // Find the parent template for each template file and create a route
   // with it.
   templatesWithoutRoot.forEach((templateFile) => {
-    let parentTemplates = filter(templatesWithoutRoot, (template) =>
+    let parentTemplates = _.filter(templatesWithoutRoot, template =>
+      // eslint-disable-next-line comma-dangle
       templateFile.requirePath.indexOf(template.file.dirname) === 0
     )
     // Sort parent templates by directory length. In cases
     // where a template has multiple parents
     // e.g. /_template.js/blog/_template.js/archive/_template.js
     // we want to nest this template under its most immediate parent.
-    parentTemplates = sortBy(parentTemplates, (template) => {
+    parentTemplates = _.sortBy(parentTemplates, (template) => {
       if (template) {
         return template.file.dirname.length
       } else {
         return 0
       }
     })
-    const parentTemplateFile = last(parentTemplates)
+    const parentTemplateFile = _.last(parentTemplates)
     let parentRoute
     if (parentTemplateFile) {
       parentRoute = templatesHash[parentTemplateFile.file.dirname]
@@ -66,7 +82,7 @@ module.exports = (files, pagesReq) => {
     // Create new route for the template.
     const route = {
       path: prefixLink(templateFile.templatePath),
-      component: pagesReq(`./${templateFile.requirePath}`),
+      component: requireComponent(pagesReq(`./${templateFile.requirePath}`)),
       childRoutes: [],
       indexRoute: {},
       pages,
@@ -82,7 +98,15 @@ module.exports = (files, pagesReq) => {
   })
 
   const staticFileTypes = [
+    'ipynb',
     'md',
+    'rmd',
+    'mkd',
+    'mkdn',
+    'mdwn',
+    'mdown',
+    'markdown',
+    'litcoffee',
     'html',
     'json',
     'yaml',
@@ -90,15 +114,23 @@ module.exports = (files, pagesReq) => {
   ]
   const reactComponentFileTypes = [
     'js',
+    'ts',
     'jsx',
+    'tsx',
     'cjsx',
   ]
   const wrappers = {}
   staticFileTypes.forEach((type) => {
     try {
+      // $FlowIssue - https://github.com/facebook/flow/issues/1975
       wrappers[type] = require(`wrappers/${type}`)
     } catch (e) {
-      // Ignore error.
+      // Ignore module not found errors; show others on console
+      if (e.code !== 'MODULE_NOT_FOUND'
+          && (e.message && !e.message.match(/^Cannot find module/))
+          && typeof console !== 'undefined') {
+        console.error('Error requiring wrapper', type, ':', e)
+      }
     }
   })
 
@@ -110,23 +142,18 @@ module.exports = (files, pagesReq) => {
       page.data = pagesReq(`./${page.requirePath}`)
     } else if (reactComponentFileTypes.indexOf(page.file.ext) !== -1) {
       handler = pagesReq(`./${page.requirePath}`)
-      page.data = (() => {
-        if (pagesReq(`./${page.requirePath}`).metadata) {
-          return pagesReq(`./${page.requirePath}`).metadata()
-        } else {
-          return {}
-        }
-      })()
+      page.data = page.data = (page.data === undefined) ? {} : page.data
     }
 
     // Determine parent template for page.
-    const parentTemplates = filter(templatesWithoutRoot, (templateFile) =>
+    const parentTemplates = _.filter(templatesWithoutRoot, templateFile =>
+      // eslint-disable-next-line comma-dangle
       page.requirePath.indexOf(templateFile.file.dirname) === 0
     )
 
-    const sortedParentTemplates = sortBy(parentTemplates, (route) => route.file.dirname.length)
+    const sortedParentTemplates = _.sortBy(parentTemplates, route => route.file.dirname.length)
 
-    const parentTemplateFile = last(sortedParentTemplates)
+    const parentTemplateFile = _.last(sortedParentTemplates)
     let parentRoute
     if (parentTemplateFile) {
       parentRoute = templatesHash[parentTemplateFile.file.dirname]
