@@ -1,6 +1,14 @@
+/* @flow weak */
 import { config } from 'config'
 import invariant from 'invariant'
-import isString from 'lodash/isString'
+import _ from 'lodash'
+
+function isDataURL (s) {
+  // Regex from https://gist.github.com/bgrins/6194623#gistcomment-1671744
+  // eslint-disable-next-line
+  const regex = /^\s*data:([a-z]+\/[a-z0-9\-\+]+(;[a-z\-]+=[a-z0-9\-]+)?)?(;base64)?,[a-z0-9!\$&',\(\)\*\+,;=\-\._~:@\/\?%\s]*\s*$/i
+  return !!s.match(regex)
+}
 
 // Function to add prefix to links.
 const prefixLink = (_link) => {
@@ -12,9 +20,18 @@ const prefixLink = (_link) => {
     You're trying to build your site with links prefixed
     but you haven't set 'linkPrefix' in your config.toml.
     `
-    invariant(isString(config.linkPrefix), invariantMessage)
+    invariant(_.isString(config.linkPrefix), invariantMessage)
 
-    return config.linkPrefix + _link
+    // When url-loader is used with the limit option, some of images are encoded to
+    // base64 data url, this check prevent such links from being prefixed.
+    if (isDataURL(_link)) {
+      return _link
+    } else {
+      if (_link.slice(0, 1) !== '/') {
+        return `${config.linkPrefix}/${_link}`
+      }
+      return `${config.linkPrefix}${_link}`
+    }
   } else {
     return _link
   }
