@@ -1,3 +1,5 @@
+/* @flow weak */
+
 import React, { PropTypes } from 'react'
 import { prefixLink } from 'gatsby-helpers'
 
@@ -12,7 +14,7 @@ console.info(defaultMessage)
 
 function HTML (props) {
   return (
-    <html>
+    <html lang="en">
       <head>
         <meta charSet="utf-8" />
         <meta httpEquiv="X-UA-Compatible" content="IE=edge" />
@@ -29,6 +31,6 @@ function HTML (props) {
   )
 }
 
-HTML.propTypes = { body: PropTypes.any }
+HTML.propTypes = { body: PropTypes.node }
 
 module.exports = HTML
