@@ -1,3 +1,4 @@
+/* @flow weak */
 import React, { PropTypes } from 'react'
 
 const defaultMessage = `
@@ -13,6 +14,6 @@ function template (props) {
   return <div>{props.children}</div>
 }
 
-template.propTypes = { children: PropTypes.any }
+template.propTypes = { children: PropTypes.node }
 
 module.exports = template
