@@ -1,12 +1,11 @@
+/* @flow weak */
 import React, { PropTypes } from 'react'
 
 module.exports = React.createClass({
   propTypes: {
     route: PropTypes.shape({
       page: PropTypes.shape({
-        data: PropTypes.shape({
-          body: PropTypes.string.isRequired,
-        }),
+        data: PropTypes.object,
       }),
     }),
   },
