@@ -1,3 +1,4 @@
+/* @flow weak */
 const toml = require('toml')
 const loaderUtils = require('loader-utils')
 const path = require('path')
@@ -14,6 +15,6 @@ module.exports = function (source) {
   value.relativePath = path.relative('.', directory)
   globPages(directory, (err, pagesData) => {
     value.pages = pagesData
-    return callback(null, `module.exports = ${JSON.stringify(value, void 0, '\t')}`)
+    return callback(null, `module.exports = ${JSON.stringify(value, null, '\t')}`)
   })
 }
