@@ -1,9 +1,9 @@
+/* @flow weak */
 import resolve from 'babel-core/lib/helpers/resolve'
 import fs from 'fs'
 import path from 'path'
 import json5 from 'json5'
-import startsWith from 'lodash/startsWith'
-import isArray from 'lodash/isArray'
+import _ from 'lodash'
 import objectAssign from 'object-assign'
 import invariant from 'invariant'
 
@@ -25,13 +25,13 @@ function defaultConfig () {
  *
  */
 function resolvePlugin (pluginName, directory, type) {
-  const gatsbyPath = path.resolve('..', '..')
+  const gatsbyPath = path.resolve(__dirname, '..', '..')
   const plugin = resolve(`babel-${type}-${pluginName}`, directory) ||
     resolve(`babel-${type}-${pluginName}`, gatsbyPath) ||
     resolve(pluginName, directory) ||
     resolve(pluginName, gatsbyPath)
 
-  const name = startsWith(pluginName, 'babel') ? pluginName : `babel-${type}-${pluginName}`
+  const name = _.startsWith(pluginName, 'babel') ? pluginName : `babel-${type}-${pluginName}`
   const pluginInvariantMessage = `
   You are trying to use a Babel plugin which Gatsby cannot find. You
   can install it using "npm install --save ${name}".
@@ -60,15 +60,15 @@ function normalizeConfig (config, directory) {
   }
 
   const presets = config.presets || []
-  presets.forEach(preset => {
+  presets.forEach((preset) => {
     normalizedConfig.presets.push(resolvePlugin(preset, directory, 'preset'))
   })
 
   const plugins = config.plugins || []
-  plugins.forEach(plugin => {
+  plugins.forEach((plugin) => {
     let normalizedPlugin
 
-    if (isArray(plugin)) {
+    if (_.isArray(plugin)) {
       normalizedPlugin = [resolvePlugin(plugin[0], directory, 'plugin'), plugin[1]]
     } else {
       normalizedPlugin = resolvePlugin(plugin, directory, 'plugin')
@@ -104,6 +104,7 @@ function findBabelrc (directory) {
  */
 function findBabelPackage (directory) {
   try {
+    // $FlowIssue - https://github.com/facebook/flow/issues/1975
     const packageJson = require(path.join(directory, 'package.json'))
     return packageJson.babel
   } catch (error) {
@@ -130,7 +131,7 @@ export default function babelConfig (program, stage) {
     babelrc.presets.unshift('react-hmre')
   }
 
-  if (!babelrc.hasOwnProperty('cacheDirectory')) {
+  if (!_.has(babelrc, 'cacheDirectory')) {
     babelrc.cacheDirectory = true
   }
 
