@@ -1,6 +1,7 @@
+/* @flow weak */
 import webpack from 'webpack'
-import webpackConfig from './webpack.config'
 import fs from 'fs'
+import webpackConfig from './webpack.config'
 
 module.exports = (program, callback) => {
   const { directory } = program
