@@ -1,8 +1,11 @@
+/* @flow weak */
+/* eslint import/imports-first: 0 */
 require('node-cjsx').transform()
 import webpack from 'webpack'
+import fs from 'fs'
 import globPages from './glob-pages'
 import webpackConfig from './webpack.config'
-import fs from 'fs'
+
 const debug = require('debug')('gatsby:html')
 
 module.exports = (program, callback) => {
@@ -10,7 +13,7 @@ module.exports = (program, callback) => {
 
   globPages(directory, (err, pages) => {
     debug('generating static HTML')
-    const routes = pages.filter((page) => page.path).map((page) => page.path)
+    const routes = pages.filter(page => page.path).map(page => page.path)
 
     // Static site generation.
     const compilerConfig = webpackConfig(program, directory, 'build-html', null, routes)
