@@ -1,8 +1,10 @@
+/* @flow */
 import slash from 'slash'
 import parsePath from 'parse-filepath'
 import urlResolver from './url-resolver'
 
-export default function pathResolver (relativePath, pageData={}) {
+
+export default function pathResolver (relativePath: string, pageData: {} = {}) {
   const data = {}
 
   data.file = parsePath(relativePath)
@@ -26,3 +28,4 @@ export default function pathResolver (relativePath, pageData={}) {
 
   return data
 }
+
