@@ -1,10 +1,11 @@
+/* @flow weak */
 import { posix as path } from 'path'
-import _, { startsWith } from 'lodash'
-import invariant from 'invariant'
+import _ from 'lodash'
 
 let rewritePath
 try {
   const gatsbyNodeConfig = path.resolve(process.cwd(), './gatsby-node')
+  // $FlowIssue - https://github.com/facebook/flow/issues/1975
   const nodeConfig = require(gatsbyNodeConfig)
   rewritePath = nodeConfig.rewritePath
 } catch (e) {
@@ -16,20 +17,27 @@ try {
 export default function pathResolver (pageData, parsedPath) {
   /**
    * Determines if a hardcoded path was given in the frontmatter of a page.
+   * Then enforces a starting and trailing slash where applicable on the url.
    */
   function hardcodedPath () {
-    if (pageData.path) {
-      const pathInvariantMessage = `
-      Hardcoded paths are relative to the website root so must be prepended with a
-      forward slash. You set the path to "${pageData.path}" in "${parsedPath.path}"
-      but it should be "/${pageData.path}"
+    if (_.isUndefined(pageData.path)) return undefined
 
-      See http://bit.ly/1qeNpdy for more.
-      `
-      invariant(pageData.path.charAt(0) === '/', pathInvariantMessage)
+    let pagePath = pageData.path
+
+    // Enforce a starting slash on all paths
+    const pathStartsWithSlash = _.startsWith(pagePath, '/')
+    if (!pathStartsWithSlash) {
+      pagePath = `/${pagePath}`
     }
 
-    return pageData.path
+    // Enforce a trailing slash on all paths
+    const pathHasExtension = path.extname(pagePath) !== ''
+    const pathEndsWithSlash = _.endsWith(pagePath, '/')
+    if (!pathEndsWithSlash && !pathHasExtension) {
+      pagePath = `${pagePath}/`
+    }
+
+    return pagePath
   }
 
   /**
@@ -59,7 +67,7 @@ export default function pathResolver (pageData, parsedPath) {
    * Returns a path for a page. If the page name starts with an underscore,
    * undefined is returned as it does not become a page.
    */
-  if (!startsWith(parsedPath.name, '_')) {
+  if (!_.startsWith(parsedPath.name, '_')) {
     return hardcodedPath() || rewrittenPath() || defaultPath()
   } else {
     return undefined
