@@ -1,3 +1,5 @@
+/* @flow weak */
+/* eslint import/imports-first: 0 */
 require('node-cjsx').transform()
 import Hapi from 'hapi'
 import Boom from 'boom'
@@ -6,18 +8,28 @@ import ReactDOMServer from 'react-dom/server'
 import webpack from 'webpack'
 import Negotiator from 'negotiator'
 import parsePath from 'parse-filepath'
-import find from 'lodash/find'
+import _ from 'lodash'
 import webpackRequire from 'webpack-require'
 import WebpackPlugin from 'hapi-webpack-plugin'
 import opn from 'opn'
 import fs from 'fs'
 import glob from 'glob'
+import chalk from 'chalk'
+import startWithDynamicPort from './startWithDynamicPort'
 
 import globPages from './glob-pages'
 import webpackConfig from './webpack.config'
 const debug = require('debug')('gatsby:application')
 
-module.exports = (program) => {
+// Display a nice noscript message in dev mode explaining that server-side rendering
+// is not enabled in develop mode
+const devNoScript = `<noscript>
+  The Gatsby development server does not work without JavaScript enabled.
+  If you'd like to test how your site works without JavaScript, first build the site
+  'gatsby build' and then serve the built site 'gatsby serve-build'
+</noscript>`
+
+function startServer (program) {
   const directory = program.directory
 
   // Load pages for the site.
@@ -45,8 +57,8 @@ module.exports = (program) => {
       const HTML = factory()
       debug('Configuring develop server')
 
-      // Setup and start Hapi to serve html + static files + webpack-hot-middleware.
       const server = new Hapi.Server()
+
       server.connection({
         host: program.host,
         port: program.port,
@@ -63,8 +75,8 @@ module.exports = (program) => {
           try {
             const htmlElement = React.createElement(
               HTML, {
-                body: '',
-              }
+                body: devNoScript,
+              },
             )
             let html = ReactDOMServer.renderToStaticMarkup(htmlElement)
             html = `<!DOCTYPE html>\n${html}`
@@ -81,7 +93,7 @@ module.exports = (program) => {
         path: '/{path*}',
         handler: {
           directory: {
-            path: `${directory}/pages`,
+            path: `${program.directory}/pages`,
             listing: false,
             index: false,
           },
@@ -93,9 +105,9 @@ module.exports = (program) => {
 
         // Try to map the url path to match an actual path of a file on disk.
         const parsed = parsePath(request.path)
-        const page = find(pages, (p) => p.path === (`${parsed.dirname}/`))
+        const page = _.find(pages, p => p.path === (`${parsed.dirname}/`))
 
-        let absolutePath = `${directory}/pages`
+        let absolutePath = `${program.directory}/pages`
         let path
         if (page) {
           path = `/${parsePath(page.requirePath).dirname}/${parsed.basename}`
@@ -119,6 +131,10 @@ module.exports = (program) => {
         } else if (request.path === '/bundle.js') {
           reply.continue()
         } else if (negotiator.mediaType() === 'text/html') {
+          // If the path does not end with a slash, add it.
+          if (request.path[request.path.length - 1] !== '/') {
+            request.path += '/' // eslint-disable-line no-param-reassign
+          }
           request.setUrl(`/html${request.path}`)
           reply.continue()
         } else {
@@ -144,7 +160,7 @@ module.exports = (program) => {
         },
       }
 
-      server.register({
+      return server.register({
         register: WebpackPlugin,
         options: {
           compiler,
@@ -160,28 +176,27 @@ module.exports = (program) => {
         server.start((e) => {
           if (e) {
             if (e.code === 'EADDRINUSE') {
-              const finder = require('process-finder')
-              finder.find({ elevate: false, port: program.port }, (startErr, pids) => {
-                const msg =
-`We were unable to start Gatsby on port ${program.port} as there's already a process
-listening on that port (PID: ${pids[0]}). You can either use a different port
-(e.g. gatsby develop --port ${parseInt(program.port, 10) + 1}) or stop the process already listening
-on your desired port.`
-                console.log(msg)
-                process.exit()
-              })
+              // eslint-disable-next-line max-len
+              console.log(chalk.red(`Unable to start Gatsby on port ${program.port} as there's already a process listing on that port.`))
             } else {
-              console.log(e)
-              process.exit()
+              console.log(chalk.red(e))
             }
+
+            process.exit()
           } else {
             if (program.open) {
               opn(server.info.uri)
             }
-            console.log('Listening at:', server.info.uri)
+            console.log(chalk.green('Server started successfully!'))
+            console.log()
+            console.log('Listening at:')
+            console.log()
+            console.log('  ', chalk.cyan(server.info.uri))
           }
         })
       })
     })
   })
 }
+
+module.exports = startWithDynamicPort(startServer)
