@@ -1,22 +1,12 @@
+/* @flow weak */
 import glob from 'glob'
 import buildPage from './build-page'
+import pageFileTypes from './page-file-types'
+
 const debug = require('debug')('gatsby:glob')
 
 function globQuery (directory) {
-  // Make this list easy to modify through the config?
-  // Or just keep adding extensions...?
-  const fileTypesToGlob = [
-    'coffee',
-    'cjsx',
-    'jsx',
-    'js',
-    'md',
-    'html',
-    'json',
-    'yaml',
-    'toml',
-  ]
-  const fileGlobQuery = fileTypesToGlob.map((type) => `*.${type}`)
+  const fileGlobQuery = pageFileTypes.map(type => `*.${type}`)
   const joinedFileQuery = fileGlobQuery.join('|')
   return `${directory}/pages/**/?(${joinedFileQuery})`
 }
@@ -25,7 +15,7 @@ let globCache
 function createCache (directory, callback) {
   const pagesData = []
 
-  glob(globQuery(directory), null, (err, pages) => {
+  glob(globQuery(directory), { follow: true }, (err, pages) => {
     if (err) { return callback(err) }
 
     pages.forEach((page) => {
