@@ -1,3 +1,4 @@
+/* @flow weak */
 import { exec } from 'child_process'
 import fs from 'fs-extra'
 import sysPath from 'path'
@@ -29,7 +30,7 @@ const install = (rootPath, callback) => {
   })
 }
 
-const ignored = (path) => !/^\.(git|hg)$/.test(sysPath.basename(path))
+const ignored = path => !/^\.(git|hg)$/.test(sysPath.basename(path))
 
 // Copy starter from file system.
 //
