@@ -1,10 +1,11 @@
+/* @flow weak */
 // This file is auto-written and used by Gatsby to require
 // files from your pages directory.
 module.exports = function (callback) {
-  let context = require.context('./pages', true)
+  let context = require.context('./pages', true, /(coffee|cjsx|ts|tsx|jsx|js|md|rmd|mkdn?|mdwn|mdown|markdown|litcoffee|ipynb|html|json|yaml|toml)$/) // eslint-disable-line
   if (module.hot) {
     module.hot.accept(context.id, () => {
-      context = require.context('./pages', true)
+      context = require.context('./pages', true, /(coffee|cjsx|ts|tsx|jsx|js|md|rmd|mkdn?|mdwn|mdown|markdown|litcoffee|ipynb|html|json|yaml|toml)$/) // eslint-disable-line
       return callback(context)
     })
   }
