@@ -1,3 +1,4 @@
+/* @flow weak */
 const logger = require('tracer').colorConsole()
 
 const initStarter = require('./init-starter')
@@ -12,6 +13,6 @@ module.exports = (rootPath, starter='gh:gatsbyjs/gatsby-starter-default') => {
       if (error) {
         logger.error(error)
       }
-    }
+    },
   )
 }
