@@ -1,12 +1,14 @@
+/* @flow weak */
 import path from 'path'
 import glob from 'glob'
 import fs from 'fs-extra'
 import async from 'async'
 import parsePath from 'parse-filepath'
-import find from 'lodash/find'
+import _ from 'lodash'
+import globPages from './glob-pages'
+
 const debug = require('debug')('gatsby:post-build')
 
-import globPages from './glob-pages'
 
 module.exports = (program, cb) => {
   const directory = program.directory
@@ -31,7 +33,7 @@ module.exports = (program, cb) => {
       }
 
       if (!(oldDirectory === '/')) {
-        const page = find(pages, (p) => {
+        const page = _.find(pages, (p) => {
           // Ignore files that start with underscore (they're not pages).
           if (p.file.name.slice(0, 1) !== '_') {
             return parsePath(p.requirePath).dirname === oldDirectory
@@ -51,14 +53,17 @@ module.exports = (program, cb) => {
       }
 
       newPath = `${directory}/public/${newPath}`
-      return fs.copy(file, newPath, (error) => callback(error))
+      return fs.copy(file, newPath, error => callback(error))
     }
 
     // Copy static assets to public folder.
-    const globString = `${directory}/pages/**/?(*.jpg|*.png|*.pdf|*.gif|*.ico|*.svg)`
-    return glob(globString, null, (e, files) =>
+    const assetTypes = '*.jpg|*.jpeg|*.png|*.pdf|*.gif|*.ico|*.svg|*.pdf|*.txt|CNAME'
+    const globString = `${directory}/pages/**/?(${assetTypes})`
+    return glob(globString, { follow: true }, (e, files) =>
       async.map(files, copy, (error, results) =>
+        // eslint-disable-next-line comma-dangle
         cb(error, results)
+      // eslint-disable-next-line comma-dangle
       )
     )
   })
