@@ -1,20 +1,21 @@
+/* @flow weak */
 import Hapi from 'hapi'
 import opn from 'opn'
+import chalk from 'chalk'
+import startWithDynamicPort from './startWithDynamicPort'
 
 const debug = require('debug')('gatsby:application')
 
-module.exports = (program) => {
+function startServer (program, launchPort) {
   const directory = program.directory
+  const serverPort = launchPort || program.port
 
   debug('Serving /public')
-
-  // Setup and start Hapi to static files.
-
   const server = new Hapi.Server()
 
   server.connection({
     host: program.host,
-    port: program.port,
+    port: serverPort,
   })
 
   server.route({
@@ -32,11 +33,25 @@ module.exports = (program) => {
 
   server.start((e) => {
     if (e) {
-      console.log(e)
+      if (e.code === 'EADDRINUSE') {
+        // eslint-disable-next-line max-len
+        console.log(chalk.red(`Unable to start Gatsby on port ${serverPort} as there's already a process listing on that port.`))
+      } else {
+        console.log(chalk.red(e))
+      }
+
+      process.exit()
+    } else {
+      if (program.open) {
+        opn(server.info.uri)
+      }
+      console.log(chalk.green('Server started successfully!'))
+      console.log()
+      console.log('Listening at:')
+      console.log()
+      console.log('  ', chalk.cyan(server.info.uri))
     }
-    if (program.open) {
-      opn(server.info.uri)
-    }
-    return console.log('Listening at:', server.info.uri)
   })
 }
+
+module.exports = startWithDynamicPort(startServer)
