@@ -1,11 +1,13 @@
+/* @flow weak */
 import React from 'react'
 import { renderToString, renderToStaticMarkup } from 'react-dom/server'
 import { match, RouterContext } from 'react-router'
 import createRoutes from 'create-routes'
 import Html from 'html'
-const loadContext = require('.gatsby-context')
 import { pages, config } from 'config'
 import { prefixLink } from '../isomorphic/gatsby-helpers'
+
+const loadContext = require('.gatsby-context')
 
 let routes
 loadContext((pagesReq) => {
@@ -29,6 +31,7 @@ module.exports = (locals, callback) => {
       }
 
       const html = `<!DOCTYPE html>\n ${renderToStaticMarkup(
+        // eslint-disable-next-line comma-dangle
         <Html body={body} {...renderProps} />
       )}`
       callback(null, html)
