@@ -1,13 +1,19 @@
 import test from 'ava'
-import fs from 'fs-extra'
+import fse from 'fs-extra'
 import Promise from 'bluebird'
-const remove = Promise.promisify(fs.remove)
-const tmpdir = require('os').tmpdir()
-
 import { gatsby } from '../support'
 
-test('calling gatsby new succesfully creates new site from default starter', async t => {
-  await remove(`${tmpdir}/gatsby-default-starter`)
-  const noArgs = await gatsby(['new', `${tmpdir}/tmp/gatsby-default-starter`])
+const fs = Promise.promisifyAll(fse)
+const tmpdir = require('os').tmpdir()
+
+
+// eslint-disable-next-line
+test('calling gatsby new succesfully creates new site from default starter', async (t) => {
+  const sitePath = `${tmpdir}/gatsby-default-starter`
+  await fs.remove(sitePath)
+  const noArgs = await gatsby(['new', sitePath])
+  const file = await fs.statAsync(`${sitePath}/html.js`)
+
   t.is(noArgs.code, 0)
+  t.truthy(file)
 })
