@@ -2,7 +2,7 @@ import fs from 'fs-extra'
 import path from 'path'
 import cheerio from 'cheerio'
 import Promise from 'bluebird'
-import { concat } from 'lodash'
+import _ from 'lodash'
 import { spawn as spawnNative } from 'child_process'
 const remove = Promise.promisify(fs.remove)
 const gatsbyCli = path.resolve('..', '..', 'lib', 'bin', 'cli.js')
@@ -13,10 +13,10 @@ export function spawn (command, args = [], options = {}) {
     let stdout = ''
     let stderr = ''
     const child = spawnNative(command, args, options)
-    child.stdout.on('data', data => { stdout += data })
-    child.stderr.on('data', data => { stderr += data })
+    child.stdout.on('data', (data) => { stdout += data })
+    child.stderr.on('data', (data) => { stderr += data })
     child.on('error', error => reject({ error, stderr, stdout }))
-    child.on('exit', code => {
+    child.on('exit', (code) => {
       if (code === 0) {
         resolve({ code, stdout, stderr })
       } else {
@@ -27,7 +27,7 @@ export function spawn (command, args = [], options = {}) {
 }
 
 export function gatsby (args = [], options = {}) {
-  const spawnArguments = concat(['--', gatsbyCli], args)
+  const spawnArguments = _.concat(['--', gatsbyCli], args)
   return spawn(babel, spawnArguments, options)
 }
 
